Properly style keyboard selection for the dual datepicker.
This change adds a more consistent experience when selecting a week
with a keyboard.

Before the change, react-datepicker added a --keyboard-selected
modifier to the day that was currently selected with a keyboard. In
the context of the weekly datepicker, having that does not make sense
and it pretty confusing. After the change, we are adding a custom
class to the parent (that is week) of the keyboard-selected day that
provides a background to the week that will be selected as a result of
keyboard selection.