Hide modal content until the backing data reaches a state that must be displayed.
Both modals depend on a RemoteData value to display their data and
transition through the states. Until now, there was no explicit
handling for the RemoteData.Requested state where we've just displayed
an empty modal. Having a fixed width hepled, but not too much, because
on fast networks we ended up seeing a flash of the empty modal.

With this change, we display the overlay but slide the modal content
offscreen such that it is not visible until the backing data
transtions to (StllLoading|Error|Available) states and we actually
have something to display. The overlay provides instant feedback for
the click action.