Fix the issue with not cleaning up after datepicker on navigation in production builds
Production build redirects away from the previous page before running
all of the clean-up commands. The existing datepicker clean-up
mechanism relied on the input element being in the DOM, which is not
the case in production builds. (and also looks like it's not always the
case in dev builds as well)

Updated datepicker cleanup mechanism stores a reference to the
datepicker instance itself instead of relying on the host input
element to be in the DOM when we need to clean-up after the page,
which makes it so it does not matter whether we've already been
redirected away from the page.